libkrun: relocate build_microvm out of the vmm crate
Move build_microvm() and the attach_* device helpers it calls from
src/vmm/src/builder.rs into src/libkrun/src/builder.rs. Port the old C
API's call site to the new location.

build_microvm's visibility drops from pub to pub(crate) since every
caller is now within the libkrun crate. The pub fn primitives
(choose_payload, create_guest_memory, setup_vm, etc.) stay in the vmm
crate and are called across the existing crate boundary.

Promote four macOS-only Vmm fields to pub (needed by libkrun's
builder.rs across the crate boundary): vm_ctl_tx, vm_ctl_rx,
paused, paused_at.
