+++ src/vmm/src/builder.rs
@@ -3,123 +3,67 @@
 
 //! Enables pre-boot setup, instantiation and booting of a Firecracker VMM.
 
+#[cfg(any(target_os = "macos", feature = "tee"))]
 use crossbeam_channel::Sender;
 #[cfg(target_os = "macos")]
 use crossbeam_channel::unbounded;
-use kernel::cmdline::Cmdline;
 #[cfg(target_os = "macos")]
 use std::collections::HashMap;
 use std::fmt::{Display, Formatter};
+#[cfg(any(
+    target_arch = "x86_64",
+    all(feature = "vhost-user", target_os = "linux")
+))]
 use std::fs::File;
-use std::io::{self, IsTerminal, Read};
+use std::io::{self, Read};
 #[cfg(unix)]
 use std::os::fd::AsRawFd;
 #[cfg(unix)]
-use std::os::fd::{BorrowedFd, FromRawFd};
-#[cfg(windows)]
-use std::os::windows::io::{AsRawHandle, BorrowedHandle, FromRawHandle};
-#[cfg(windows)]
-use std::path::PathBuf;
-use std::sync::atomic::AtomicI32;
+use std::os::fd::BorrowedFd;
+#[cfg(all(feature = "vhost-user", target_os = "linux"))]
+use std::os::fd::FromRawFd;
 use std::sync::{Arc, Mutex};
 #[cfg(windows)]
 use utils::windows::SendHandle;
-#[cfg(windows)]
-use windows_sys::Win32::Foundation::INVALID_HANDLE_VALUE;
-#[cfg(windows)]
-use windows_sys::Win32::System::Console::{
-    GetStdHandle, STD_ERROR_HANDLE, STD_INPUT_HANDLE, STD_OUTPUT_HANDLE,
-};
 
 use super::{Error, Vmm};
 
 #[cfg(target_arch = "x86_64")]
 use crate::device_manager::legacy::PortIODeviceManager;
 use crate::device_manager::mmio::MMIODeviceManager;
-use crate::resources::{
-    DefaultVirtioConsoleConfig, PortConfig, TsiFlags, VirtioConsoleConfigMode, VmResources,
-};
+use crate::resources::VmResources;
 use crate::vmm_config::external_kernel::{ExternalKernel, KernelFormat};
-#[cfg(feature = "net")]
-use crate::vmm_config::net::NetBuilder;
+use devices::legacy::IrqChip;
 #[cfg(target_arch = "x86_64")]
-use devices::legacy::Cmos;
+use devices::legacy::IrqChipDevice;
 #[cfg(all(target_os = "linux", target_arch = "riscv64"))]
 use devices::legacy::KvmAia;
-#[cfg(target_arch = "x86_64")]
-use devices::legacy::KvmIoapic;
 use devices::legacy::Serial;
 #[cfg(target_os = "macos")]
 use devices::legacy::VcpuList;
-#[cfg(target_os = "macos")]
-use devices::legacy::{GicV3, HvfGicV3};
-#[cfg(target_arch = "x86_64")]
-use devices::legacy::{IoApic, IrqChipT};
-use devices::legacy::{IrqChip, IrqChipDevice};
-#[cfg(all(target_os = "linux", target_arch = "aarch64"))]
-use devices::legacy::{KvmGicV2, KvmGicV3};
-#[cfg(not(any(feature = "tee", feature = "aws-nitro")))]
-use devices::virtio::passthrough::PermissionSemantics;
-use devices::virtio::{MmioTransport, PortDescription, VirtioDevice, Vsock, port_io};
+
+use devices::virtio::{MmioTransport, VirtioDevice};
 
 #[cfg(feature = "tee")]
 use kbs_types::Tee;
 
 use crate::device_manager;
-#[cfg(feature = "tdx")]
-use crate::linux::tee::tdshim::{self, TdShim};
-#[cfg(all(feature = "vhost-user", target_os = "linux"))]
-use crate::resources::VhostUserDeviceConfig;
-#[cfg(target_os = "linux")]
-use crate::signal_handler::register_sigint_handler;
-#[cfg(target_os = "linux")]
-use crate::signal_handler::register_sigwinch_handler;
 use crate::terminal::{term_restore_mode, term_set_raw_mode};
-#[cfg(feature = "blk")]
-use crate::vmm_config::block::BlockBuilder;
-#[cfg(feature = "tdx")]
-use crate::vmm_config::firmware::TeeFirmwareType;
-#[cfg(not(any(feature = "tee", feature = "aws-nitro")))]
-use crate::vmm_config::fs::FsDeviceConfig;
-use crate::vmm_config::kernel_cmdline::DEFAULT_KERNEL_CMDLINE;
 #[cfg(target_os = "linux")]
 use crate::vstate::KvmContext;
-#[cfg(all(target_os = "linux", feature = "tee"))]
-use crate::vstate::MeasuredRegion;
 use crate::vstate::{Error as VstateError, Vcpu, VcpuConfig, Vm};
 use arch::{ArchMemoryInfo, InitrdConfig};
 use device_manager::shm::ShmManager;
-#[cfg(any(feature = "gpu", feature = "vhost-user"))]
-use devices::display::DisplayInfo;
-#[cfg(feature = "gpu")]
-use devices::display::NoopDisplayBackend;
-#[cfg(not(any(feature = "tee", feature = "aws-nitro")))]
-use devices::virtio::{VirtioShmRegion, fs::ExportTable};
 use flate2::read::GzDecoder;
-#[cfg(feature = "gpu")]
-use krun_display::DisplayBackend;
-#[cfg(feature = "gpu")]
-use krun_display::IntoDisplayBackend;
-#[cfg(feature = "amd-sev")]
-use kvm_bindings::KVM_MAX_CPUID_ENTRIES;
-#[cfg(unix)]
-use libc::{STDERR_FILENO, STDIN_FILENO, STDOUT_FILENO};
 #[cfg(target_arch = "x86_64")]
 use linux_loader::loader::{self, KernelLoader};
-#[cfg(unix)]
-use nix::unistd::isatty;
 use polly::event_manager::{Error as EventManagerError, EventManager};
 use utils::eventfd::EventFd;
+#[cfg(feature = "tee")]
 use utils::worker_message::WorkerMessage;
-#[cfg(not(any(feature = "tee", feature = "aws-nitro")))]
-use vm_memory::Address;
 use vm_memory::Bytes;
 #[cfg(all(feature = "vhost-user", target_os = "linux"))]
 use vm_memory::FileOffset;
-#[cfg(not(feature = "aws-nitro"))]
-use vm_memory::GuestMemory;
-#[cfg(feature = "tdx")]
-use vm_memory::GuestMemoryRegion;
 #[cfg(all(target_arch = "x86_64", not(feature = "tee")))]
 use vm_memory::GuestRegionMmap;
 #[cfg(all(target_arch = "x86_64", not(feature = "tee")))]
@@ -593,823 +537,6 @@ pub fn choose_payload(vm_resources: &VmResources) -> Result<Payload, StartMicrov
     } else {
         Err(StartMicrovmError::MissingKernelConfig)
     }
-}
-
-#[cfg(feature = "tdx")]
-fn measure_tdshim_regions(
-    td_shim: TdShim,
-    vm_resources: &super::resources::VmResources,
-    guest_memory: &GuestMemoryMmap,
-) -> Result<(Vec<MeasuredRegion>, u64), StartMicrovmError> {
-    td_shim
-        .load_sections(guest_memory)
-        .map_err(|e| StartMicrovmError::TdShimError(format!("{e}")))?;
-
-    let high_fw = td_shim.high_firmware_range();
-    let ram_regions: Vec<(u64, u64)> = guest_memory
-        .iter()
-        .filter_map(|region| {
-            let start = region.start_addr().0;
-            let len = region.len();
-            if let Some((fw_start, fw_end)) = high_fw
-                && start >= fw_start
-                && start < fw_end
-            {
-                return None;
-            }
-            Some((start, len))
-        })
-        .collect();
-
-    let startup_64 = vm_resources
-        .kernel_bundle
-        .as_ref()
-        .ok_or(StartMicrovmError::MissingKernelConfig)?
-        .entry_addr;
-    // When an initrd is present, the HOB entry point is backed up by
-    // TRAMPOLINE_SIZE so td-shim lands on the boot_params trampoline
-    // that patches initrd address/size before jumping to startup_64.
-    let hob_entry_point = if vm_resources.initrd_bundle.is_some() {
-        startup_64 - tdshim::TRAMPOLINE_SIZE
-    } else {
-        startup_64
-    };
-    td_shim
-        .generate_hobs(guest_memory, hob_entry_point, &ram_regions)
-        .map_err(|e| StartMicrovmError::TdShimError(format!("{e}")))?;
-
-    // All RAM as one block (attributes=0, add but don't measure), plus the
-    // high firmware sections (BFV etc.) with their per-section attributes.
-    // Low-address TDVF sections (TempMem, TD_HOB) fall inside the RAM range
-    // and must not be added separately — TDX rejects duplicate TDH.MEM.PAGE.ADD.
-    let mut regions: Vec<MeasuredRegion> = guest_memory
-        .iter()
-        .filter(|r| r.start_addr().0 < arch::x86_64::layout::MMIO_MEM_START)
-        .map(|r| MeasuredRegion {
-            guest_addr: r.start_addr().0,
-            host_addr: guest_memory.get_host_address(r.start_addr()).unwrap() as u64,
-            size: r.len() as usize,
-            attributes: 0,
-        })
-        .collect();
-
-    for section in &td_shim.sections {
-        if section.memory_address >= arch::x86_64::layout::MMIO_MEM_START {
-            regions.push(MeasuredRegion {
-                guest_addr: section.memory_address,
-                host_addr: guest_memory
-                    .get_host_address(GuestAddress(section.memory_address))
-                    .unwrap() as u64,
-                size: section.memory_data_size as usize,
-                attributes: section.attributes,
-            });
-        }
-    }
-
-    Ok((regions, td_shim.hob_address))
-}
-
-#[cfg(feature = "tdx")]
-fn measure_qboot_regions(
-    vm_resources: &super::resources::VmResources,
-    guest_memory: &GuestMemoryMmap,
-) -> Result<(Vec<MeasuredRegion>, u64), StartMicrovmError> {
-    let qboot_size = if let Some(qboot_bundle) = &vm_resources.qboot_bundle {
-        qboot_bundle.size
-    } else {
-        return Err(StartMicrovmError::MissingKernelConfig);
-    };
-    let regions = vec![
-        MeasuredRegion {
-            guest_addr: 0,
-            host_addr: guest_memory.get_host_address(GuestAddress(0)).unwrap() as u64,
-            size: 0x8000_0000,
-            attributes: 0,
-        },
-        MeasuredRegion {
-            guest_addr: arch::FIRMWARE_START,
-            host_addr: guest_memory
-                .get_host_address(GuestAddress(arch::FIRMWARE_START))
-                .unwrap() as u64,
-            size: qboot_size,
-            attributes: 1,
-        },
-    ];
-    Ok((regions, 0u64))
-}
-
-/// Builds and starts a microVM based on the current Firecracker VmResources configuration.
-///
-/// This is the default build recipe, one could build other microVM flavors by using the
-/// independent functions in this module instead of calling this recipe.
-///
-/// An `Arc` reference of the built `Vmm` is also plugged in the `EventManager`, while another
-/// is returned.
-pub fn build_microvm(
-    vm_resources: &super::resources::VmResources,
-    event_manager: &mut EventManager,
-    _shutdown_efd: Option<EventFd>,
-    _sender: Sender<WorkerMessage>,
-) -> std::result::Result<Arc<Mutex<Vmm>>, StartMicrovmError> {
-    let payload = choose_payload(vm_resources)?;
-
-    #[cfg(feature = "tdx")]
-    let td_shim_parsed = match &vm_resources.tee_firmware_config {
-        Some(tee_fw_cfg) => match tee_fw_cfg.fw_type {
-            TeeFirmwareType::TdShim => Some(
-                TdShim::parse(&tee_fw_cfg.path)
-                    .map_err(|e| StartMicrovmError::TdShimError(format!("{e}")))?,
-            ),
-        },
-        None => None,
-    };
-
-    #[cfg(feature = "tdx")]
-    let fw_range_for_mem = td_shim_parsed.as_ref().and_then(|ts| {
-        ts.high_firmware_range()
-            .map(|(start, end)| (start, (end - start) as usize))
-    });
-    #[cfg(all(feature = "tee", not(feature = "tdx")))]
-    let fw_range_for_mem: Option<(u64, usize)> = None;
-
-    #[cfg(not(feature = "tee"))]
-    let fs_shm_sizes: Vec<Option<usize>> = vm_resources.fs.iter().map(|f| f.shm_size).collect();
-    #[cfg(feature = "tee")]
-    let fs_shm_sizes: Vec<Option<usize>> = Vec::new();
-
-    let (guest_memory, arch_memory_info, mut _shm_manager, payload_config) = create_guest_memory(
-        vm_resources
-            .vm_config()
-            .mem_size_mib
-            .ok_or(StartMicrovmError::MissingMemSizeConfig)?,
-        vm_resources.kernel_bundle.as_ref(),
-        #[cfg(feature = "tee")]
-        vm_resources.qboot_bundle.as_ref(),
-        #[cfg(feature = "tee")]
-        vm_resources.initrd_bundle.as_ref(),
-        vm_resources.firmware_config.as_ref(),
-        &fs_shm_sizes,
-        vm_resources
-            .gpu_virgl_flags
-            .map(|_| vm_resources.gpu_shm_size.unwrap_or(1 << 33)),
-        {
-            #[cfg(all(feature = "vhost-user", target_os = "linux"))]
-            {
-                !vm_resources.vhost_user_devices.is_empty()
-            }
-            #[cfg(not(all(feature = "vhost-user", target_os = "linux")))]
-            {
-                false
-            }
-        },
-        &payload,
-        #[cfg(feature = "tee")]
-        fw_range_for_mem,
-    )?;
-
-    let vcpu_config = vm_resources.vcpu_config();
-
-    // Clone the command-line so that a failed boot doesn't pollute the original.
-    #[allow(unused_mut)]
-    let mut kernel_cmdline = Cmdline::new(arch::CMDLINE_MAX_SIZE);
-    if let Some(cmdline) = payload_config.kernel_cmdline {
-        kernel_cmdline.insert_str(cmdline.as_str()).unwrap();
-    } else if let Some(cmdline) = &vm_resources.kernel_cmdline.prolog {
-        kernel_cmdline.insert_str(cmdline).unwrap();
-    } else {
-        kernel_cmdline.insert_str(DEFAULT_KERNEL_CMDLINE).unwrap();
-    }
-
-    if let Some(cmdline) = &vm_resources.kernel_cmdline.krun_env {
-        kernel_cmdline.insert_str(cmdline.as_str()).unwrap();
-    }
-
-    if let Some(kernel_console) = &vm_resources.kernel_console {
-        let cmdline = kernel_cmdline.as_str();
-        let console_start_idx = cmdline.find("console=").unwrap();
-        let console_end_idx = cmdline
-            .get(console_start_idx..)
-            .and_then(|s| s.find(" ").map(|i| i + console_start_idx));
-
-        let cmdline = cmdline.replace(
-            &cmdline[console_start_idx..console_end_idx.unwrap()],
-            format!("console={kernel_console}").as_str(),
-        );
-        kernel_cmdline = Cmdline::new(arch::CMDLINE_MAX_SIZE);
-        kernel_cmdline.insert_str(cmdline).unwrap();
-    }
-
-    // Write the TD-Shim initrd trampoline and firmware sections into guest memory
-    // BEFORE setup_vm()/memory_init(). At this point the GuestMemoryMmap is backed
-    // by plain anonymous mmap pages. Once KVM registers the memory slots (memory_init),
-    // pages become KVM_MEM_PRIVATE and TDH.MEM.PAGE.ADD copies the shared content to
-    // the TD's private memory — so any writes here are guaranteed to reach the TD.
-    #[cfg(feature = "tdx")]
-    if let Some(ref td_shim) = td_shim_parsed {
-        td_shim
-            .load_sections(&guest_memory)
-            .map_err(|e| StartMicrovmError::TdShimError(format!("{e}")))?;
-
-        arch::x86_64::setup_mptable_for_tdshim(
-            &guest_memory,
-            vm_resources.vm_config().vcpu_count.unwrap_or(1),
-        )
-        .map_err(|e| StartMicrovmError::Internal(Error::ConfigureSystem(e)))?;
-
-        // Place the trampoline INSIDE the kernel image, just before startup_64.
-        // The HOB entry_point must be in the kernel's address range — any address
-        // outside (like 0x801000 in the TD_HOB section) causes td-shim to treat
-        // the jump target as HOB data and behave incorrectly before the kernel runs.
-        //
-        // The bytes before startup_64 are PE/COFF header data that td-shim's
-        // RawVmLinux boot path never reads or executes, so overwriting them is safe.
-        //
-        // td-shim creates its own boot_params with cmd_line_ptr=0, so the
-        // kernel falls back to CONFIG_CMDLINE and we have no way to inject a
-        // custom cmdline.  Patching cmd_line_ptr here lets load_cmdline's
-        // content (written to CMDLINE_START) reach the kernel.
-        //
-        // Note: the trampoline is only written when an initrd is present. Without
-        // an initrd no trampoline is placed, so boot_params.cmd_line_ptr will remain
-        // 0 and the kernel won't see any cmdline passed via load_cmdline. This is a
-        // known limitation of the current TD-Shim boot path.
-        if let (Some(kernel_bundle), Some(initrd_bundle)) =
-            (&vm_resources.kernel_bundle, &vm_resources.initrd_bundle)
-        {
-            let trampoline_addr = kernel_bundle.entry_addr - tdshim::TRAMPOLINE_SIZE;
-            let trampoline = tdshim::build_boot_params_trampoline(
-                arch::x86_64::layout::INITRD_SEV_START as u32,
-                initrd_bundle.size.try_into().unwrap(),
-                arch::x86_64::layout::CMDLINE_START as u32,
-            );
-            guest_memory
-                .write(&trampoline, GuestAddress(trampoline_addr))
-                .map_err(|e| StartMicrovmError::TdShimError(format!("{e}")))?;
-        }
-    }
-
-    #[cfg(not(feature = "tee"))]
-    #[allow(unused_mut)]
-    let mut vm = setup_vm(&guest_memory, vm_resources.nested_enabled)?;
-
-    #[cfg(feature = "tee")]
-    let (_kvm, vm) = {
-        let kvm = KvmContext::new()
-            .map_err(Error::KvmContext)
-            .map_err(StartMicrovmError::Internal)?;
-        let vm = setup_vm(
-            &kvm,
-            &guest_memory,
-            vm_resources,
-            #[cfg(feature = "tdx")]
-            _sender.clone(),
-        )?;
-        (kvm, vm)
-    };
-
-    #[cfg(feature = "tee")]
-    let tee = vm_resources.tee_config().tee;
-
-    #[cfg(feature = "amd-sev")]
-    let snp_launcher = match tee {
-        Tee::Snp => Some(
-            vm.snp_secure_virt_prepare(&guest_memory)
-                .map_err(StartMicrovmError::SecureVirtPrepare)?,
-        ),
-        _ => None,
-    };
-
-    #[cfg(feature = "tdx")]
-    let mut tdx_launcher = match tee {
-        Tee::Tdx => vm
-            .tdx_secure_virt_prepare()
-            .map_err(StartMicrovmError::SecureVirtPrepare)?,
-        _ => panic!(),
-    };
-
-    #[cfg(all(feature = "tee", not(feature = "tdx")))]
-    let measured_regions = {
-        println!("Injecting and measuring memory regions. This may take a while.");
-
-        let qboot_size = if let Some(qboot_bundle) = &vm_resources.qboot_bundle {
-            qboot_bundle.size
-        } else {
-            return Err(StartMicrovmError::MissingKernelConfig);
-        };
-        let (kernel_guest_addr, kernel_size) =
-            if let Some(kernel_bundle) = &vm_resources.kernel_bundle {
-                (kernel_bundle.guest_addr, kernel_bundle.size)
-            } else {
-                return Err(StartMicrovmError::MissingKernelConfig);
-            };
-        let (initrd_addr, initrd_size) = if let Some(initrd_config) = &payload_config.initrd_config
-        {
-            (initrd_config.address, initrd_config.size)
-        } else {
-            return Err(StartMicrovmError::MissingKernelConfig);
-        };
-
-        vec![
-            MeasuredRegion {
-                guest_addr: arch::FIRMWARE_START,
-                host_addr: guest_memory
-                    .get_host_address(GuestAddress(arch::FIRMWARE_START))
-                    .unwrap() as u64,
-                size: qboot_size,
-                attributes: 0,
-            },
-            MeasuredRegion {
-                guest_addr: kernel_guest_addr,
-                host_addr: guest_memory
-                    .get_host_address(GuestAddress(kernel_guest_addr))
-                    .unwrap() as u64,
-                size: kernel_size,
-                attributes: 0,
-            },
-            MeasuredRegion {
-                guest_addr: initrd_addr.0,
-                host_addr: guest_memory.get_host_address(initrd_addr).unwrap() as u64,
-                size: initrd_size,
-                attributes: 0,
-            },
-            MeasuredRegion {
-                guest_addr: arch::x86_64::layout::ZERO_PAGE_START,
-                host_addr: guest_memory
-                    .get_host_address(GuestAddress(arch::x86_64::layout::ZERO_PAGE_START))
-                    .unwrap() as u64,
-                size: 4096,
-                attributes: 0,
-            },
-        ]
-    };
-
-    #[cfg(feature = "tdx")]
-    let (measured_regions, tdx_hob_address) = {
-        println!("Injecting and measuring memory regions. This may take a while.");
-
-        if let Some(td_shim) = td_shim_parsed {
-            measure_tdshim_regions(td_shim, vm_resources, &guest_memory)?
-        } else {
-            measure_qboot_regions(vm_resources, &guest_memory)?
-        }
-    };
-
-    let mut serial_devices = Vec::new();
-
-    // We can't call to `setup_terminal_raw_mode` until `Vmm` is created,
-    // so let's keep track of FDs connected to legacy serial devices here
-    // and set raw mode on them later.
-    let mut serial_ttys = Vec::new();
-
-    #[cfg(unix)]
-    for s in &vm_resources.serial_consoles {
-        let input: Option<Box<dyn devices::legacy::ReadableFd + Send>> = if s.input_fd >= 0 {
-            let file = unsafe { File::from_raw_fd(s.input_fd) };
-            if file.is_terminal() {
-                serial_ttys.push(unsafe { BorrowedFd::borrow_raw(file.as_raw_fd()) });
-            }
-            Some(Box::new(file))
-        } else {
-            None
-        };
-
-        let output: Option<Box<dyn io::Write + Send>> = if s.output_fd >= 0 {
-            Some(Box::new(unsafe { File::from_raw_fd(s.output_fd) }))
-        } else {
-            None
-        };
-
-        serial_devices.push(setup_serial_device(event_manager, input, output)?);
-    }
-
-    #[cfg(windows)]
-    for s in &vm_resources.serial_consoles {
-        let input: Option<Box<dyn devices::legacy::ReadableFd + Send>> =
-            if is_valid_handle(s.input_handle.as_raw_handle()) {
-                if unsafe {
-                    BorrowedHandle::borrow_raw(s.input_handle.as_raw_handle()).is_terminal()
-                } {
-                    serial_ttys.push(s.input_handle);
-                }
-                Some(Box::new(unsafe {
-                    File::from_raw_handle(s.input_handle.as_raw_handle())
-                }))
-            } else {
-                None
-            };
-
-        let output: Option<Box<dyn io::Write + Send>> =
-            if is_valid_handle(s.output_handle.as_raw_handle()) {
-                Some(Box::new(unsafe {
-                    File::from_raw_handle(s.output_handle.as_raw_handle())
-                }))
-            } else {
-                None
-            };
-
-        serial_devices.push(setup_serial_device(event_manager, input, output)?);
-    }
-
-    let exit_evt = EventFd::new(utils::eventfd::EFD_NONBLOCK)
-        .map_err(Error::EventFd)
-        .map_err(StartMicrovmError::Internal)?;
-
-    #[cfg(target_arch = "x86_64")]
-    // Safe to unwrap 'serial_device' as it's always 'Some' on x86_64.
-    // x86_64 uses the i8042 reset event as the Vmm exit event.
-    let mut pio_device_manager = PortIODeviceManager::new(
-        Arc::new(Mutex::new(Cmos::new(
-            arch_memory_info.ram_below_gap,
-            arch_memory_info.ram_above_gap,
-        ))),
-        serial_devices,
-        exit_evt
-            .try_clone()
-            .map_err(Error::EventFd)
-            .map_err(StartMicrovmError::Internal)?,
-    )
-    .map_err(Error::CreateLegacyDevice)
-    .map_err(StartMicrovmError::Internal)?;
-
-    // Instantiate the MMIO device manager.
-    // 'mmio_base' address has to be an address which is protected by the kernel
-    // and is architectural specific.
-    #[allow(unused_mut)]
-    let mut mmio_device_manager = MMIODeviceManager::new(
-        &mut (arch::MMIO_MEM_START.clone()),
-        (arch::IRQ_BASE, arch::IRQ_MAX),
-    );
-
-    #[cfg(target_os = "macos")]
-    let vcpu_list = {
-        let cpu_count = vm_resources.vm_config().vcpu_count.unwrap();
-        Arc::new(VcpuList::new(cpu_count as u64))
-    };
-
-    let vcpus;
-    let intc: IrqChip;
-    // For x86_64 we need to create the interrupt controller before calling `KVM_CREATE_VCPUS`
-    // while on aarch64 we need to do it the other way around.
-    #[cfg(target_arch = "x86_64")]
-    {
-        let ioapic: Box<dyn IrqChipT> = if vm_resources.split_irqchip {
-            Box::new(
-                IoApic::new(vm.fd(), _sender.clone())
-                    .map_err(StartMicrovmError::CreateKvmIrqChip)?,
-            )
-        } else {
-            Box::new(KvmIoapic::new(vm.fd()).map_err(StartMicrovmError::CreateKvmIrqChip)?)
-        };
-        intc = Arc::new(Mutex::new(IrqChipDevice::new(ioapic)));
-
-        attach_legacy_devices(
-            &vm,
-            vm_resources.split_irqchip,
-            &mut pio_device_manager,
-            &mut mmio_device_manager,
-            Some(intc.clone()),
-        )?;
-
-        let kernel_boot = vm_resources.firmware_config.is_none() && !cfg!(feature = "tee");
-
-        vcpus = create_vcpus_x86_64(
-            &vm,
-            &vcpu_config,
-            &guest_memory,
-            payload_config.entry_addr,
-            &pio_device_manager.io_bus,
-            &exit_evt,
-            kernel_boot,
-            payload_config.pvh,
-            #[cfg(feature = "tee")]
-            _sender,
-        )
-        .map_err(StartMicrovmError::Internal)?;
-    }
-
-    #[cfg(feature = "tdx")]
-    {
-        for vcpu in &vcpus {
-            vcpu.tdx_secure_virt_prepare(&mut tdx_launcher);
-        }
-        vm.tdx_secure_virt_init_vcpus(&mut tdx_launcher, tdx_hob_address)
-            .unwrap();
-    }
-
-    // On aarch64, the vCPUs need to be created (i.e call KVM_CREATE_VCPU) and configured before
-    // setting up the IRQ chip because the `KVM_CREATE_VCPU` ioctl will return error if the IRQCHIP
-    // was already initialized.
-    // Search for `kvm_arch_vcpu_create` in arch/arm/kvm/arm.c.
-    #[cfg(all(target_arch = "aarch64", target_os = "linux"))]
-    {
-        vcpus = create_vcpus_aarch64(
-            &vm,
-            &vcpu_config,
-            &arch_memory_info,
-            payload_config.entry_addr,
-            &exit_evt,
-        )
-        .map_err(StartMicrovmError::Internal)?;
-
-        intc = {
-            // The SoC in some popular boards (namely, the RPi family) doesn't support an
-            // architected vGIC, which is required for requesting KVM the instantiation of a
-            // GICv3. To relieve the users from having to configure the gic version manually,
-            // try first to instantiate a GICv3, and fall back to a GICv2 if it fails.
-            let vcpu_count = vm_resources.vm_config().vcpu_count.unwrap() as u64;
-            let gic = match KvmGicV3::new(vm.fd(), vcpu_count) {
-                Ok(gicv3) => IrqChipDevice::new(Box::new(gicv3)),
-                Err(_) => {
-                    warn!("KVM GICv3 creation failed, falling back to KVM GICv2");
-                    IrqChipDevice::new(Box::new(KvmGicV2::new(vm.fd(), vcpu_count)))
-                }
-            };
-            Arc::new(Mutex::new(gic))
-        };
-
-        attach_legacy_devices(
-            &vm,
-            &mut mmio_device_manager,
-            &mut kernel_cmdline,
-            intc.clone(),
-            serial_devices,
-        )?;
-    }
-
-    #[cfg(all(target_arch = "aarch64", target_os = "macos"))]
-    {
-        intc = {
-            // If the system supports the in-kernel GIC, use it. Otherwise, fall back to the
-            // userspace implementation.
-            let gic = match HvfGicV3::new(vm_resources.vm_config().vcpu_count.unwrap() as u64) {
-                Ok(hvfgic) => IrqChipDevice::new(Box::new(hvfgic)),
-                Err(_) => IrqChipDevice::new(Box::new(GicV3::new(vcpu_list.clone()))),
-            };
-            Arc::new(Mutex::new(gic))
-        };
-
-        vcpus = create_vcpus_aarch64(
-            &vm,
-            &vcpu_config,
-            &arch_memory_info,
-            payload_config.entry_addr,
-            &exit_evt,
-            vcpu_list.clone(),
-            vm_resources.nested_enabled,
-        )
-        .map_err(StartMicrovmError::Internal)?;
-
-        attach_legacy_devices(
-            &vm,
-            &mut mmio_device_manager,
-            &mut kernel_cmdline,
-            intc.clone(),
-            serial_devices,
-            event_manager,
-            _shutdown_efd,
-        )?;
-    }
-
-    #[cfg(all(target_arch = "riscv64", target_os = "linux"))]
-    {
-        vcpus = create_vcpus_riscv64(
-            &vm,
-            &vcpu_config,
-            &guest_memory,
-            payload_config.entry_addr,
-            &exit_evt,
-        )
-        .map_err(StartMicrovmError::Internal)?;
-
-        intc = Arc::new(Mutex::new(IrqChipDevice::new(Box::new(
-            KvmAia::new(vm.fd(), vm_resources.vm_config().vcpu_count.unwrap() as u32).unwrap(),
-        ))));
-
-        attach_legacy_devices(
-            &vm,
-            &mut mmio_device_manager,
-            &mut kernel_cmdline,
-            intc.clone(),
-            serial_devices,
-        )?;
-    }
-
-    // We use this atomic to record the exit code set by init/init.c in the VM.
-    let exit_code = Arc::new(AtomicI32::new(i32::MAX));
-
-    #[cfg(target_os = "macos")]
-    let (vm_ctl_tx, vm_ctl_rx) = utils::pollable_channel::pollable_channel()
-        .map_err(Error::EventFd)
-        .map_err(StartMicrovmError::Internal)?;
-
-    let mut vmm = Vmm {
-        guest_memory,
-        arch_memory_info,
-        kernel_cmdline,
-        vcpus_handles: Vec::new(),
-        exit_evt,
-        exit_observers: Vec::new(),
-        exit_code: exit_code.clone(),
-        vm,
-        mmio_device_manager,
-        #[cfg(target_arch = "x86_64")]
-        pio_device_manager,
-        #[cfg(target_os = "macos")]
-        vm_ctl_tx,
-        #[cfg(target_os = "macos")]
-        vm_ctl_rx,
-        #[cfg(target_os = "macos")]
-        paused: false,
-        #[cfg(target_os = "macos")]
-        paused_at: 0,
-    };
-
-    // Set raw mode for FDs that are connected to legacy serial devices.
-    for serial_tty in serial_ttys {
-        setup_terminal_raw_mode(&mut vmm, Some(serial_tty), false);
-    }
-
-    #[cfg(not(feature = "tee"))]
-    attach_balloon_device(&mut vmm, event_manager, intc.clone())?;
-    #[cfg(not(feature = "tee"))]
-    {
-        #[cfg(all(feature = "vhost-user", target_os = "linux"))]
-        {
-            const VIRTIO_ID_RNG: u32 = 4;
-            for (index, device_config) in vm_resources.vhost_user_devices.iter().enumerate() {
-                attach_vhost_user_device(
-                    &mut vmm,
-                    &_shm_manager,
-                    event_manager,
-                    intc.clone(),
-                    device_config,
-                    index,
-                    vm_resources.displays.first().cloned(),
-                )?;
-            }
-
-            let has_vhost_user_rng = vm_resources
-                .vhost_user_devices
-                .iter()
-                .any(|dev| dev.device_type == VIRTIO_ID_RNG);
-
-            if !has_vhost_user_rng {
-                attach_rng_device(&mut vmm, event_manager, intc.clone())?;
-            }
-        }
-
-        #[cfg(not(all(feature = "vhost-user", target_os = "linux")))]
-        {
-            attach_rng_device(&mut vmm, event_manager, intc.clone())?;
-        }
-    }
-    for (console_id, console_cfg) in vm_resources.virtio_consoles.iter().enumerate() {
-        attach_console_devices(
-            &mut vmm,
-            event_manager,
-            intc.clone(),
-            vm_resources,
-            Some(console_cfg),
-            console_id as u32,
-        )?;
-    }
-
-    #[cfg(not(any(feature = "tee", feature = "aws-nitro")))]
-    let export_table: Option<ExportTable> = if cfg!(feature = "gpu") {
-        Some(Default::default())
-    } else {
-        None
-    };
-
-    #[cfg(feature = "gpu")]
-    if let Some(virgl_flags) = vm_resources.gpu_virgl_flags {
-        let display_backend = vm_resources
-            .display_backend
-            .unwrap_or_else(|| NoopDisplayBackend::into_display_backend(None));
-
-        attach_gpu_device(
-            &mut vmm,
-            &mut _shm_manager,
-            #[cfg(not(feature = "tee"))]
-            export_table.clone(),
-            intc.clone(),
-            virgl_flags,
-            Box::from(&vm_resources.displays[..]),
-            display_backend,
-            #[cfg(target_os = "macos")]
-            _sender.clone(),
-        )?;
-    }
-
-    #[cfg(feature = "input")]
-    if !vm_resources.input_backends.is_empty() {
-        attach_input_devices(&mut vmm, &vm_resources.input_backends, intc.clone())?;
-    }
-
-    #[cfg(not(any(feature = "tee", feature = "aws-nitro")))]
-    attach_fs_devices(
-        &mut vmm,
-        &vm_resources.fs,
-        &mut _shm_manager,
-        #[cfg(not(feature = "tee"))]
-        export_table,
-        intc.clone(),
-        exit_code,
-        #[cfg(target_os = "macos")]
-        _sender,
-    )?;
-    #[cfg(feature = "blk")]
-    attach_block_devices(&mut vmm, &vm_resources.block, intc.clone())?;
-
-    if let Some(vsock) = vm_resources.vsock.get() {
-        attach_unixsock_vsock_device(&mut vmm, vsock, event_manager, intc.clone())?;
-        let tsi_flags = vm_resources.vsock.tsi_flags();
-        if tsi_flags.contains(TsiFlags::HIJACK_INET) {
-            vmm.kernel_cmdline.insert_str("tsi_hijack")?;
-        }
-        if tsi_flags.contains(TsiFlags::HIJACK_UNIX) {
-            vmm.kernel_cmdline.insert_str("tsi_hijack_unix")?;
-        }
-    }
-
-    #[cfg(feature = "net")]
-    attach_net_devices(&mut vmm, &vm_resources.net, intc.clone())?;
-    #[cfg(feature = "net")]
-    if vm_resources.dhcp_client {
-        vmm.kernel_cmdline.insert_str("KRUN_DHCP=1")?;
-    }
-
-    if let Some(s) = &vm_resources.kernel_cmdline.epilog {
-        vmm.kernel_cmdline.insert_str(s).unwrap();
-    };
-
-    // Write the kernel command line to guest memory. This is x86_64 specific, since on
-    // aarch64 the command line will be specified through the FDT.
-    // For the TD-Shim path, the cmdline is written so TD-Shim can reference it when
-    // populating boot_params for the Linux kernel (cmd_line_ptr already points here
-    // via configure_system).
-    #[cfg(all(target_arch = "x86_64", not(feature = "tee")))]
-    load_cmdline(&vmm)?;
-    #[cfg(all(target_arch = "x86_64", feature = "tdx"))]
-    if vm_resources.tee_firmware_config.is_some() {
-        load_cmdline(&vmm)?;
-    }
-
-    vmm.configure_system(
-        vcpus.as_slice(),
-        &intc,
-        &payload_config.initrd_config,
-        &vm_resources.smbios_oem_strings,
-        payload_config.pvh,
-    )
-    .map_err(StartMicrovmError::Internal)?;
-
-    #[cfg(feature = "tee")]
-    {
-        match tee {
-            #[cfg(feature = "amd-sev")]
-            Tee::Snp => {
-                let cpuid = _kvm
-                    .fd()
-                    .get_supported_cpuid(KVM_MAX_CPUID_ENTRIES)
-                    .map_err(VstateError::KvmCpuId)
-                    .map_err(StartMicrovmError::SecureVirtAttest)?;
-                vmm.kvm_vm()
-                    .snp_secure_virt_measure(
-                        cpuid,
-                        vmm.guest_memory(),
-                        measured_regions,
-                        snp_launcher.unwrap(),
-                    )
-                    .map_err(StartMicrovmError::SecureVirtAttest)?;
-            }
-            #[cfg(feature = "tdx")]
-            Tee::Tdx => {
-                vmm.kvm_vm()
-                    .tdx_secure_virt_prepare_memory(&mut tdx_launcher, &measured_regions)
-                    .unwrap();
-                vmm.kvm_vm()
-                    .tdx_secure_virt_finalize_vm(tdx_launcher)
-                    .map_err(StartMicrovmError::SecureVirtPrepare)?;
-            }
-            _ => return Err(StartMicrovmError::InvalidTee),
-        }
-
-        println!("Starting TEE/microVM.");
-    }
-
-    vmm.start_vcpus(vcpus)
-        .map_err(StartMicrovmError::Internal)?;
-
-    // Clippy thinks we don't need Arc<Mutex<...
-    // but we don't want to change the event_manager interface
-    #[allow(clippy::arc_with_non_send_sync)]
-    let vmm = Arc::new(Mutex::new(vmm));
-    event_manager
-        .add_subscriber(vmm.clone())
-        .map_err(StartMicrovmError::RegisterEvent)?;
-
-    Ok(vmm)
 }
 
 fn load_external_kernel(
@@ -2370,266 +1497,6 @@ pub fn attach_mmio_device(
     Ok(())
 }
 
-#[cfg(not(any(feature = "tee", feature = "aws-nitro")))]
-fn attach_fs_devices(
-    vmm: &mut Vmm,
-    fs_devs: &[FsDeviceConfig],
-    shm_manager: &mut ShmManager,
-    #[cfg(not(feature = "tee"))] export_table: Option<ExportTable>,
-    intc: IrqChip,
-    exit_code: Arc<AtomicI32>,
-    #[cfg(target_os = "macos")] map_sender: Sender<WorkerMessage>,
-) -> std::result::Result<(), StartMicrovmError> {
-    use self::StartMicrovmError::*;
-
-    for (i, config) in fs_devs.iter().enumerate() {
-        let fs = Arc::new(Mutex::new(
-            devices::virtio::Fs::new(
-                config.fs_id.clone(),
-                PermissionSemantics::LinuxComplete,
-                config.shared_dir.clone(),
-                exit_code.clone(),
-                config.read_only,
-                config.virtual_entries.clone(),
-            )
-            .unwrap(),
-        ));
-
-        let id = format!("{}{}", String::from(fs.lock().unwrap().id()), i);
-
-        if let Some(shm_region) = shm_manager.fs_region(i) {
-            fs.lock().unwrap().set_shm_region(VirtioShmRegion {
-                host_addr: vmm
-                    .guest_memory
-                    .get_host_address(shm_region.guest_addr)
-                    .map_err(StartMicrovmError::ShmHostAddr)? as u64,
-                guest_addr: shm_region.guest_addr.raw_value(),
-                size: shm_region.size,
-            });
-        }
-
-        #[cfg(not(feature = "tee"))]
-        if let Some(export_table) = export_table.as_ref() {
-            fs.lock().unwrap().set_export_table(export_table.clone());
-        }
-
-        #[cfg(target_os = "macos")]
-        fs.lock().unwrap().set_map_sender(map_sender.clone());
-
-        // The device mutex mustn't be locked here otherwise it will deadlock.
-        attach_mmio_device(vmm, id, intc.clone(), fs).map_err(RegisterFsDevice)?;
-    }
-
-    Ok(())
-}
-
-#[cfg(unix)]
-fn autoconfigure_console_ports(
-    vmm: &mut Vmm,
-    _vm_resources: &VmResources,
-    cfg: Option<&DefaultVirtioConsoleConfig>,
-) -> std::result::Result<Vec<PortDescription>, StartMicrovmError> {
-    let (input_fd, output_fd, err_fd) = match cfg {
-        Some(c) => (c.input_fd, c.output_fd, c.err_fd),
-        None => (STDIN_FILENO, STDOUT_FILENO, STDERR_FILENO),
-    };
-    {
-        let input_is_terminal =
-            input_fd >= 0 && isatty(unsafe { BorrowedFd::borrow_raw(input_fd) }).unwrap_or(false);
-        let output_is_terminal =
-            output_fd >= 0 && isatty(unsafe { BorrowedFd::borrow_raw(output_fd) }).unwrap_or(false);
-        let error_is_terminal =
-            err_fd >= 0 && isatty(unsafe { BorrowedFd::borrow_raw(err_fd) }).unwrap_or(false);
-
-        let term_fd = if input_is_terminal {
-            Some(unsafe { BorrowedFd::borrow_raw(input_fd) })
-        } else if output_is_terminal {
-            Some(unsafe { BorrowedFd::borrow_raw(output_fd) })
-        } else if error_is_terminal {
-            Some(unsafe { BorrowedFd::borrow_raw(err_fd) })
-        } else {
-            None
-        };
-
-        let forwarding_sigint;
-        let console_input = if input_is_terminal && input_fd >= 0 {
-            forwarding_sigint = false;
-            Some(port_io::input_to_raw_fd_dup(input_fd).unwrap())
-        } else {
-            #[cfg(target_os = "linux")]
-            {
-                forwarding_sigint = true;
-                let sigint_input = port_io::PortInputSigInt::new();
-                let sigint_input_fd = sigint_input.sigint_evt().as_raw_fd();
-                register_sigint_handler(sigint_input_fd)
-                    .map_err(StartMicrovmError::RegisterFsSigwinch)?;
-                Some(Box::new(sigint_input) as _)
-            }
-            #[cfg(not(target_os = "linux"))]
-            {
-                forwarding_sigint = false;
-                Some(port_io::input_empty().unwrap())
-            }
-        };
-
-        let console_output = if output_is_terminal && output_fd >= 0 {
-            Some(port_io::output_to_raw_fd_dup(output_fd).unwrap())
-        } else {
-            Some(port_io::output_to_log_as_err())
-        };
-
-        let terminal_properties = term_fd
-            .map(|fd| port_io::term_fd(fd.as_raw_fd()).unwrap())
-            .unwrap_or_else(|| port_io::term_fixed_size(0, 0));
-
-        setup_terminal_raw_mode(vmm, term_fd, forwarding_sigint);
-
-        let mut ports = vec![PortDescription::console(
-            console_input,
-            console_output,
-            terminal_properties,
-        )];
-
-        if input_fd >= 0 && !input_is_terminal {
-            ports.push(PortDescription::input_pipe(
-                "krun-stdin",
-                port_io::input_to_raw_fd_dup(input_fd).unwrap(),
-            ));
-        }
-
-        if output_fd >= 0 && !output_is_terminal {
-            ports.push(PortDescription::output_pipe(
-                "krun-stdout",
-                port_io::output_to_raw_fd_dup(output_fd).unwrap(),
-            ));
-        };
-
-        if err_fd >= 0 && !error_is_terminal {
-            ports.push(PortDescription::output_pipe(
-                "krun-stderr",
-                port_io::output_to_raw_fd_dup(err_fd).unwrap(),
-            ));
-        }
-
-        Ok(ports)
-    }
-}
-
-#[cfg(windows)]
-fn is_valid_handle(h: *mut core::ffi::c_void) -> bool {
-    !h.is_null() && h != INVALID_HANDLE_VALUE
-}
-
-#[cfg(target_os = "windows")]
-fn autoconfigure_console_ports(
-    vmm: &mut Vmm,
-    vm_resources: &VmResources,
-    cfg: Option<&DefaultVirtioConsoleConfig>,
-    creating_implicit_console: bool,
-) -> std::result::Result<Vec<PortDescription>, StartMicrovmError> {
-    use self::StartMicrovmError::*;
-
-    let mut console_output_path: Option<PathBuf> = None;
-    if let Some(path) = vm_resources.console_output.clone() {
-        if !vm_resources.disable_implicit_console && creating_implicit_console {
-            console_output_path = Some(path)
-        }
-    }
-
-    if let Some(console_output_path) = console_output_path {
-        let file = File::create(console_output_path).map_err(OpenConsoleFile)?;
-        // Manually emulate our Legacy behavior: In the case of output_path we have always used the
-        // stdin to determine the console size
-        let stdin_h = unsafe { BorrowedHandle::borrow_raw(GetStdHandle(STD_INPUT_HANDLE)) };
-        let term_h = if stdin_h.is_terminal() {
-            port_io::term_handle(stdin_h.as_raw_handle()).unwrap()
-        } else {
-            port_io::term_fixed_size(0, 0)
-        };
-        Ok(vec![PortDescription::console(
-            Some(port_io::input_empty().unwrap()),
-            Some(port_io::output_file(file).unwrap()),
-            term_h,
-        )])
-    } else {
-        let (input_h, output_h, err_h) = match cfg {
-            Some(c) => (
-                c.input_handle.as_raw_handle(),
-                c.output_handle.as_raw_handle(),
-                c.err_handle.as_raw_handle(),
-            ),
-            None => unsafe {
-                (
-                    GetStdHandle(STD_INPUT_HANDLE),
-                    GetStdHandle(STD_OUTPUT_HANDLE),
-                    GetStdHandle(STD_ERROR_HANDLE),
-                )
-            },
-        };
-        let input_is_terminal = (unsafe { BorrowedHandle::borrow_raw(input_h) }).is_terminal();
-        let output_is_terminal = (unsafe { BorrowedHandle::borrow_raw(output_h) }).is_terminal();
-        let error_is_terminal = (unsafe { BorrowedHandle::borrow_raw(err_h) }).is_terminal();
-
-        let term_h = if input_is_terminal {
-            Some(SendHandle::new(input_h))
-        } else if output_is_terminal {
-            Some(SendHandle::new(output_h))
-        } else if error_is_terminal {
-            Some(SendHandle::new(err_h))
-        } else {
-            None
-        };
-
-        let forwarding_sigint = false;
-        let console_input = if input_is_terminal {
-            Some(port_io::input_to_handle_dup(input_h).unwrap())
-        } else {
-            Some(port_io::input_empty().unwrap())
-        };
-
-        let console_output = if output_is_terminal {
-            Some(port_io::output_to_handle_dup(output_h).unwrap())
-        } else {
-            Some(port_io::output_to_log_as_err())
-        };
-
-        let terminal_properties = term_h
-            .map(|h| port_io::term_handle(h.as_raw_handle()).unwrap())
-            .unwrap_or_else(|| port_io::term_fixed_size(0, 0));
-
-        setup_terminal_raw_mode(vmm, term_h, forwarding_sigint);
-
-        let mut ports = vec![PortDescription::console(
-            console_input,
-            console_output,
-            terminal_properties,
-        )];
-
-        if is_valid_handle(input_h) && !input_is_terminal {
-            ports.push(PortDescription::input_pipe(
-                "krun-stdin",
-                port_io::input_to_handle_dup(input_h).unwrap(),
-            ));
-        }
-
-        if is_valid_handle(output_h) && !output_is_terminal {
-            ports.push(PortDescription::output_pipe(
-                "krun-stdout",
-                port_io::output_to_handle_dup(output_h).unwrap(),
-            ));
-        };
-
-        if is_valid_handle(err_h) && !error_is_terminal {
-            ports.push(PortDescription::output_pipe(
-                "krun-stderr",
-                port_io::output_to_handle_dup(err_h).unwrap(),
-            ));
-        }
-
-        Ok(ports)
-    }
-}
-
 #[cfg(unix)]
 pub fn setup_terminal_raw_mode(
     vmm: &mut Vmm,
@@ -2677,519 +1544,3 @@ pub fn setup_terminal_raw_mode(
     }
 }
 
-#[cfg(unix)]
-fn create_explicit_ports(
-    vmm: &mut Vmm,
-    port_configs: &[PortConfig],
-) -> std::result::Result<Vec<PortDescription>, StartMicrovmError> {
-    let mut ports = Vec::with_capacity(port_configs.len());
-
-    for port_cfg in port_configs {
-        let port_desc = match port_cfg {
-            PortConfig::Tty { name, tty_fd } => {
-                assert!(*tty_fd > 0, "PortConfig::Tty must have a valid tty_fd");
-                let term_fd = unsafe { BorrowedFd::borrow_raw(*tty_fd) };
-                setup_terminal_raw_mode(vmm, Some(term_fd), false);
-
-                PortDescription {
-                    name: name.clone().into(),
-                    input: Some(port_io::input_to_raw_fd_dup(*tty_fd).unwrap()),
-                    output: Some(port_io::output_to_raw_fd_dup(*tty_fd).unwrap()),
-                    terminal: Some(port_io::term_fd(*tty_fd).unwrap()),
-                }
-            }
-            PortConfig::InOut {
-                name,
-                input_fd,
-                output_fd,
-            } => PortDescription {
-                name: name.clone().into(),
-                input: if *input_fd < 0 {
-                    None
-                } else {
-                    Some(port_io::input_to_raw_fd_dup(*input_fd).unwrap())
-                },
-                output: if *output_fd < 0 {
-                    None
-                } else {
-                    Some(port_io::output_to_raw_fd_dup(*output_fd).unwrap())
-                },
-                terminal: None,
-            },
-        };
-
-        ports.push(port_desc);
-    }
-
-    Ok(ports)
-}
-
-#[cfg(target_os = "windows")]
-fn create_explicit_ports(
-    vmm: &mut Vmm,
-    port_configs: &[PortConfig],
-) -> std::result::Result<Vec<PortDescription>, StartMicrovmError> {
-    let mut ports = Vec::with_capacity(port_configs.len());
-
-    for port_cfg in port_configs {
-        let port_desc = match port_cfg {
-            PortConfig::Tty { name, tty_handle } => {
-                assert!(
-                    is_valid_handle(tty_handle.as_raw_handle()),
-                    "PortConfig::Tty must have a valid tty_handle"
-                );
-                let term_h = SendHandle::new(tty_handle.as_raw_handle());
-                setup_terminal_raw_mode(vmm, Some(term_h), false);
-
-                PortDescription {
-                    name: name.clone().into(),
-                    input: Some(port_io::input_to_handle_dup(tty_handle.as_raw_handle()).unwrap()),
-                    output: Some(
-                        port_io::output_to_handle_dup(tty_handle.as_raw_handle()).unwrap(),
-                    ),
-                    terminal: Some(port_io::term_handle(tty_handle.as_raw_handle()).unwrap()),
-                }
-            }
-            PortConfig::InOut {
-                name,
-                input_handle,
-                output_handle,
-            } => PortDescription {
-                name: name.clone().into(),
-                input: if !is_valid_handle(input_handle.as_raw_handle()) {
-                    None
-                } else {
-                    Some(port_io::input_to_handle_dup(input_handle.as_raw_handle()).unwrap())
-                },
-                output: if !is_valid_handle(output_handle.as_raw_handle()) {
-                    None
-                } else {
-                    Some(port_io::output_to_handle_dup(output_handle.as_raw_handle()).unwrap())
-                },
-                terminal: None,
-            },
-        };
-
-        ports.push(port_desc);
-    }
-
-    Ok(ports)
-}
-
-fn attach_console_devices(
-    vmm: &mut Vmm,
-    event_manager: &mut EventManager,
-    intc: IrqChip,
-    vm_resources: &VmResources,
-    cfg: Option<&VirtioConsoleConfigMode>,
-    id_number: u32,
-) -> std::result::Result<(), StartMicrovmError> {
-    use self::StartMicrovmError::*;
-
-    let ports = match cfg {
-        None => autoconfigure_console_ports(vmm, vm_resources, None)?,
-        Some(VirtioConsoleConfigMode::Autoconfigure(autocfg)) => {
-            autoconfigure_console_ports(vmm, vm_resources, Some(autocfg))?
-        }
-        Some(VirtioConsoleConfigMode::Explicit(ports)) => create_explicit_ports(vmm, ports)?,
-    };
-
-    let console = Arc::new(Mutex::new(devices::virtio::Console::new(ports).unwrap()));
-
-    vmm.exit_observers.push(console.clone());
-
-    event_manager
-        .add_subscriber(console.clone())
-        .map_err(RegisterEvent)?;
-
-    #[cfg(target_os = "linux")]
-    register_sigwinch_handler(console.lock().unwrap().get_sigwinch_fd())
-        .map_err(RegisterFsSigwinch)?;
-
-    // The device mutex mustn't be locked here otherwise it will deadlock.
-    attach_mmio_device(vmm, format!("hvc{id_number}"), intc, console)
-        .map_err(RegisterConsoleDevice)?;
-
-    Ok(())
-}
-
-#[cfg(feature = "net")]
-fn attach_net_devices(
-    vmm: &mut Vmm,
-    net_devices: &NetBuilder,
-    intc: IrqChip,
-) -> Result<(), StartMicrovmError> {
-    for net_device in net_devices.list.iter() {
-        let id = net_device.lock().unwrap().id().to_string();
-
-        attach_mmio_device(vmm, id, intc.clone(), net_device.clone())
-            .map_err(StartMicrovmError::RegisterNetDevice)?;
-    }
-    Ok(())
-}
-
-fn attach_unixsock_vsock_device(
-    vmm: &mut Vmm,
-    unix_vsock: &Arc<Mutex<Vsock>>,
-    event_manager: &mut EventManager,
-    intc: IrqChip,
-) -> std::result::Result<(), StartMicrovmError> {
-    use self::StartMicrovmError::*;
-
-    event_manager
-        .add_subscriber(unix_vsock.clone())
-        .map_err(RegisterEvent)?;
-
-    let id = String::from(unix_vsock.lock().unwrap().id());
-
-    // The device mutex mustn't be locked here otherwise it will deadlock.
-    attach_mmio_device(vmm, id, intc, unix_vsock.clone()).map_err(RegisterVsockDevice)?;
-
-    Ok(())
-}
-
-#[cfg(not(feature = "tee"))]
-fn attach_balloon_device(
-    vmm: &mut Vmm,
-    event_manager: &mut EventManager,
-    intc: IrqChip,
-) -> std::result::Result<(), StartMicrovmError> {
-    use self::StartMicrovmError::*;
-
-    let balloon = Arc::new(Mutex::new(devices::virtio::Balloon::new().unwrap()));
-
-    event_manager
-        .add_subscriber(balloon.clone())
-        .map_err(RegisterEvent)?;
-
-    let id = String::from(balloon.lock().unwrap().id());
-
-    // The device mutex mustn't be locked here otherwise it will deadlock.
-    attach_mmio_device(vmm, id, intc.clone(), balloon).map_err(RegisterBalloonDevice)?;
-
-    Ok(())
-}
-
-#[cfg(feature = "blk")]
-fn attach_block_devices(
-    vmm: &mut Vmm,
-    block_devs: &BlockBuilder,
-    intc: IrqChip,
-) -> std::result::Result<(), StartMicrovmError> {
-    use self::StartMicrovmError::*;
-
-    for block in block_devs.list.iter() {
-        let id = String::from(block.lock().unwrap().id());
-
-        // The device mutex mustn't be locked here otherwise it will deadlock.
-        attach_mmio_device(vmm, id, intc.clone(), block.clone()).map_err(RegisterBlockDevice)?;
-    }
-
-    Ok(())
-}
-
-#[cfg(not(feature = "tee"))]
-fn attach_rng_device(
-    vmm: &mut Vmm,
-    event_manager: &mut EventManager,
-    intc: IrqChip,
-) -> std::result::Result<(), StartMicrovmError> {
-    use self::StartMicrovmError::*;
-
-    let rng = Arc::new(Mutex::new(devices::virtio::Rng::new().unwrap()));
-
-    event_manager
-        .add_subscriber(rng.clone())
-        .map_err(RegisterEvent)?;
-
-    let id = String::from(rng.lock().unwrap().id());
-
-    // The device mutex mustn't be locked here otherwise it will deadlock.
-    attach_mmio_device(vmm, id, intc.clone(), rng).map_err(RegisterRngDevice)?;
-
-    Ok(())
-}
-
-#[cfg(not(feature = "tee"))]
-#[cfg(all(feature = "vhost-user", target_os = "linux"))]
-fn attach_vhost_user_device(
-    vmm: &mut Vmm,
-    shm_manager: &ShmManager,
-    event_manager: &mut EventManager,
-    intc: IrqChip,
-    device_config: &VhostUserDeviceConfig,
-    device_index: usize,
-    gpu_display: Option<DisplayInfo>,
-) -> std::result::Result<(), StartMicrovmError> {
-    use self::StartMicrovmError::*;
-
-    let device_name = device_config
-        .name
-        .clone()
-        .unwrap_or_else(|| format!("vhost-user-{}", device_config.device_type));
-
-    let device = Arc::new(Mutex::new(
-        devices::virtio::VhostUserDevice::new(
-            &device_config.socket_path,
-            device_config.device_type,
-            device_name.clone(),
-            device_config.num_queues,
-            &device_config.queue_sizes,
-            gpu_display,
-        )
-        .map_err(|e| RegisterVhostUserDevice(device_manager::mmio::Error::VhostUserDevice(e)))?,
-    ));
-
-    if let Some(shm_region) = shm_manager.vhost_user_region(device_index) {
-        device.lock().unwrap().set_shm_region(VirtioShmRegion {
-            host_addr: vmm
-                .guest_memory
-                .get_host_address(shm_region.guest_addr)
-                .map_err(StartMicrovmError::ShmHostAddr)? as u64,
-            guest_addr: shm_region.guest_addr.raw_value(),
-            size: shm_region.size,
-        });
-    }
-
-    event_manager
-        .add_subscriber(device.clone())
-        .map_err(RegisterEvent)?;
-
-    attach_mmio_device(vmm, device_name, intc.clone(), device).map_err(RegisterVhostUserDevice)?;
-
-    Ok(())
-}
-
-#[cfg(feature = "gpu")]
-#[allow(clippy::too_many_arguments)]
-fn attach_gpu_device(
-    vmm: &mut Vmm,
-    shm_manager: &mut ShmManager,
-    #[cfg(not(feature = "tee"))] mut export_table: Option<ExportTable>,
-    intc: IrqChip,
-    virgl_flags: u32,
-    displays: Box<[DisplayInfo]>,
-    display_backend: DisplayBackend<'static>,
-    #[cfg(target_os = "macos")] map_sender: Sender<WorkerMessage>,
-) -> std::result::Result<(), StartMicrovmError> {
-    use self::StartMicrovmError::*;
-
-    let gpu = Arc::new(Mutex::new(
-        devices::virtio::Gpu::new(
-            virgl_flags,
-            displays,
-            display_backend,
-            #[cfg(target_os = "macos")]
-            map_sender,
-        )
-        .unwrap(),
-    ));
-
-    let id = String::from(gpu.lock().unwrap().id());
-
-    if let Some(shm_region) = shm_manager.gpu_region() {
-        gpu.lock().unwrap().set_shm_region(VirtioShmRegion {
-            host_addr: vmm
-                .guest_memory
-                .get_host_address(shm_region.guest_addr)
-                .map_err(StartMicrovmError::ShmHostAddr)? as u64,
-            guest_addr: shm_region.guest_addr.raw_value(),
-            size: shm_region.size,
-        });
-    }
-
-    #[cfg(not(feature = "tee"))]
-    if let Some(export_table) = export_table.take() {
-        gpu.lock().unwrap().set_export_table(export_table);
-    }
-
-    // The device mutex mustn't be locked here otherwise it will deadlock.
-    attach_mmio_device(vmm, id, intc, gpu).map_err(RegisterGpuDevice)?;
-
-    Ok(())
-}
-
-#[cfg(feature = "input")]
-fn attach_input_devices(
-    vmm: &mut Vmm,
-    input_backends: &[(
-        krun_input::InputConfigBackend<'static>,
-        krun_input::InputEventProviderBackend<'static>,
-    )],
-    intc: IrqChip,
-) -> std::result::Result<(), StartMicrovmError> {
-    use self::StartMicrovmError::*;
-
-    for (index, (config_backend, events_backend)) in input_backends.iter().enumerate() {
-        let input_device = Arc::new(Mutex::new(
-            devices::virtio::input::Input::new(*config_backend, *events_backend).unwrap(),
-        ));
-
-        let id = format!("input{}", index);
-        attach_mmio_device(vmm, id, intc.clone(), input_device).map_err(RegisterInputDevice)?;
-    }
-
-    Ok(())
-}
-
-#[cfg(test)]
-pub mod tests {
-    use super::*;
-    use crate::vmm_config::kernel_bundle::KernelBundle;
-
-    #[allow(unused)]
-    fn default_guest_memory(
-        mem_size_mib: usize,
-    ) -> std::result::Result<
-        (GuestMemoryMmap, ArchMemoryInfo, ShmManager, PayloadConfig),
-        StartMicrovmError,
-    > {
-        let kernel_bundle = KernelBundle {
-            host_addr: 0x1000,
-            guest_addr: 0x1000,
-            entry_addr: 0x1000,
-            size: 0x1000,
-        };
-
-        create_guest_memory(
-            mem_size_mib,
-            Some(&kernel_bundle),
-            #[cfg(feature = "tee")]
-            None,
-            #[cfg(feature = "tee")]
-            None,
-            None,
-            &[],
-            None,
-            false,
-            &Payload::Empty,
-            #[cfg(feature = "tee")]
-            None,
-        )
-    }
-
-    #[test]
-    #[cfg(target_arch = "x86_64")]
-    fn test_create_vcpus_x86_64() {
-        let vcpu_count = 2;
-
-        let vcpu_config = VcpuConfig {
-            vcpu_count,
-            ht_enabled: false,
-            cpu_template: None,
-            nested_enabled: false,
-        };
-
-        let (guest_memory, _arch_memory_info, _shm_manager, _payload_config) =
-            default_guest_memory(128).unwrap();
-        let vm = setup_vm(&guest_memory, false).unwrap();
-        let _kvmioapic = KvmIoapic::new(vm.fd()).unwrap();
-
-        // Dummy entry_addr, vcpus will not boot.
-        let entry_addr = GuestAddress(0);
-        let bus = devices::Bus::new();
-        let vcpu_vec = create_vcpus_x86_64(
-            &vm,
-            &vcpu_config,
-            &guest_memory,
-            entry_addr,
-            &bus,
-            &EventFd::new(utils::eventfd::EFD_NONBLOCK).unwrap(),
-            true,
-            false,
-        )
-        .unwrap();
-        assert_eq!(vcpu_vec.len(), vcpu_count as usize);
-    }
-
-    #[test]
-    #[cfg(all(target_arch = "aarch64", target_os = "linux"))]
-    fn test_create_vcpus_aarch64() {
-        let (guest_memory, arch_memory_info, _shm_manager, _payload_config) =
-            default_guest_memory(128).unwrap();
-        let vm = setup_vm(&guest_memory, false).unwrap();
-        let vcpu_count = 2;
-
-        let vcpu_config = VcpuConfig {
-            vcpu_count,
-            ht_enabled: false,
-            cpu_template: None,
-            nested_enabled: false,
-        };
-
-        // Dummy entry_addr, vcpus will not boot.
-        let entry_addr = GuestAddress(0);
-        let vcpu_vec = create_vcpus_aarch64(
-            &vm,
-            &vcpu_config,
-            &arch_memory_info,
-            entry_addr,
-            &EventFd::new(utils::eventfd::EFD_NONBLOCK).unwrap(),
-        )
-        .unwrap();
-        assert_eq!(vcpu_vec.len(), vcpu_count as usize);
-    }
-
-    #[test]
-    fn test_error_messages() {
-        use crate::builder::StartMicrovmError::*;
-        let err = AttachBlockDevice(io::Error::from_raw_os_error(0));
-        let _ = format!("{err}{err:?}");
-
-        let err = CreateRateLimiter(io::Error::from_raw_os_error(0));
-        let _ = format!("{err}{err:?}");
-
-        let err = Internal(Error::Serial(io::Error::from_raw_os_error(0)));
-        let _ = format!("{err}{err:?}");
-
-        let err = InvalidKernelBundle(vm_memory::mmap::MmapRegionError::InvalidPointer);
-        let _ = format!("{err}{err:?}");
-
-        let err = KernelCmdline(String::from("dummy --cmdline"));
-        let _ = format!("{err}{err:?}");
-
-        let err = LoadCommandline(kernel::cmdline::Error::TooLarge);
-        let _ = format!("{err}{err:?}");
-
-        let err = MicroVMAlreadyRunning;
-        let _ = format!("{err}{err:?}");
-
-        let err = MissingKernelConfig;
-        let _ = format!("{err}{err:?}");
-
-        let err = MissingMemSizeConfig;
-        let _ = format!("{err}{err:?}");
-
-        let err = NetDeviceNotConfigured;
-        let _ = format!("{err}{err:?}");
-
-        let err = OpenBlockDevice(io::Error::from_raw_os_error(0));
-        let _ = format!("{err}{err:?}");
-
-        let err = RegisterBlockDevice(device_manager::mmio::Error::EventFd(
-            io::Error::from_raw_os_error(0),
-        ));
-        let _ = format!("{err}{err:?}");
-
-        let err = RegisterEvent(EventManagerError::EpollCreate(
-            io::Error::from_raw_os_error(0),
-        ));
-        let _ = format!("{err}{err:?}");
-
-        let err = RegisterNetDevice(device_manager::mmio::Error::EventFd(
-            io::Error::from_raw_os_error(0),
-        ));
-        let _ = format!("{err}{err:?}");
-
-        let err = RegisterVsockDevice(device_manager::mmio::Error::EventFd(
-            io::Error::from_raw_os_error(0),
-        ));
-        let _ = format!("{err}{err:?}");
-    }
-
-    #[test]
-    fn test_kernel_cmdline_err_to_startuvm_err() {
-        let err = StartMicrovmError::from(kernel::cmdline::Error::HasSpace);
-        let _ = format!("{err}{err:?}");
-    }
-}
